вынес функции отрисовки фрагментов пункта списка и фото за итератор, функцию переводчик в утилитарные
